+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - **array:** add rotate matrix ([cbeb58a](https://github.com/air-supply94/js-algorithm/commit/cbeb58ab6cf44f0f027975ce08e19cab46b1dc98))
 - **avlTree:** 重构此类 ([bd35c86](https://github.com/air-supply94/js-algorithm/commit/bd35c8633f67618017b574312f63328c52a3930b))
 - **AvlTree:** 增加类型 ([57971b0](https://github.com/air-supply94/js-algorithm/commit/57971b0acd7f85c9633708e0c045031c7c415cb7))
+- **binarySearch:** add minEatingSpeed ([0d0ce3b](https://github.com/air-supply94/js-algorithm/commit/0d0ce3b75e4a22bbbe91e63fbcfb7888d8bc9344))
 - **binarySearchNode:** 更改遍历 ([1fae22a](https://github.com/air-supply94/js-algorithm/commit/1fae22a8cff9fc8b9e9d012f35c137b7b6d80038))
 - **binarySearchNode:** 重构该类 ([4f2d1cb](https://github.com/air-supply94/js-algorithm/commit/4f2d1cba7cdb5c8aba537194b1921de4c76405a7))
 - **BinaryTreeNode:** 去除 implements ([e4c9f4f](https://github.com/air-supply94/js-algorithm/commit/e4c9f4f19bc79e13b32c56e9222ae3c3941a55ae))